// src/api/openapi/backend/models/Vehicle.ts
export interface Vehicle {
    /**
     * 
     * @type {number}
     * @memberof Vehicle
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Vehicle
     */
    registrationPlate: string;
    /**
     * 
     * @type {Customer}
     * @memberof Vehicle
     */
    customer: Customer;
    /**
     * 
     * @type {string}
     * @memberof Vehicle
     */
    vin: string | null;
    /**
     * 
     * @type {string}
     * @memberof Vehicle
     */
    engineCode: string | null;
    /**
     * 
     * @type {string}
     * @memberof Vehicle
     */
    fuelType: string | null;
    /**
     * 
     * @type {number}
     * @memberof Vehicle
     */
    enginePower: number | null;
    /**
     * 
     * @type {number}
     * @memberof Vehicle
     */
    engineVolume: number | null;
    /**
     * 
     * @type {number}
     * @memberof Vehicle
     */
    batteryCapacity: number | null;
    /**
     * 
     * @type {string}
     * @memberof Vehicle
     */
    brand: string | null;
    /**
     * 
     * @type {string}
     * @memberof Vehicle
     */
    model: string | null;
    /**
     * 
     * @type {number}
     * @memberof Vehicle
     */
    yearOfManufacture: number | null;
    /**
     * 
     * @type {string}
     * @memberof Vehicle
     */
    created: string;
    /**
     * 
     * @type {string}
     * @memberof Vehicle
     */
    modified: string | null;
}